[Refactor][Elementwise] Let the backend own bool storage, and ask it for capability
Bool operands run on a uint8 kernel. The op did the reinterpretation: view the
input as uint8, call, view the result back as bool — nine branches across
bitwise, logical, comparison and masked_fill, all keyed off a `bool_storage`
flag the entry carried for the purpose. That is an op implementing one
backend's storage convention. A backend with native bool would have to be
wrapped in a TileLang-shaped adapter to fit.

The reinterpretation moves into the kernels that require it. Both
`_Uint8Storage{Unary,Binary}Kernel` and the two masked_fill kernels now accept
bool and return bool, which is what their docstrings already claimed. Callers
pass semantic tensors and `KernelEntry` loses `bool_storage`: the record states
semantics only, and `compute_dtype` already says what the kernel runs on.

Capability questions also went to the wrong object. Four sites read
`SUPPORTED_DTYPES` off the family default kernel class, so a `kernel_map`
override supporting a different dtype set was validated against the class it
replaced. They now ask `_selected_kernel_cls()` — the kernel that will
actually run.

Also fixes `RoundFwdOp`'s non-zero decimals path, which answered without a
kernel and so never bound the call metadata: a fresh instance left
`self.dtype` unset and `total_memory` claiming no forward had run, and after a
float32 call a float16 call still reported float32. Binding moves out of
`_entry` into `_note_call`, which every successful forward path uses. Adds the
sequence as a test, plus a `torch.compile` regression test for integer
`reciprocal`, whose manifest-driven output promotion had no committed coverage.

Removes remaining prose and examples that instruct callers to pass a `dtype=`
argument no constructor accepts.

+++ tileops/ops/norm/rms_norm.py
@@ -32,7 +32,7 @@ class RMSNormFwdOp(Op):
         tune: Whether to autotune (default ``False``).
 
     Example:
-        >>> op = RMSNormFwdOp(normalized_shape=(4096,), dtype=torch.float16)
+        >>> op = RMSNormFwdOp(normalized_shape=(4096,))
         >>> x = torch.randn(1024, 4096, dtype=torch.float16, device="cuda")
         >>> w = torch.randn(4096, dtype=torch.float16, device="cuda")
         >>> y = op(x, w)  # shape: (1024, 4096)
